put the ApolloClient setup in App because we want to await for the apollo queries to run for persisting cach
